Move CommonExtractor into tests directory
